refactor(tasks): split tasks page module boundaries
Refs #238
