+++ front/src/pages/tasks/hooks/useTaskDetailsActions.ts
@@ -3,7 +3,7 @@ import toast from 'react-hot-toast'
 import type { TaskDetails as TaskDetailsType } from '@/shared/types'
 import { useTaskDetails } from '@/pages/tasks/hooks/useTaskDetails'
 
-export const useTaskActions = (task: TaskDetailsType | undefined) => {
+export const useTaskDetailsActions = (task: TaskDetailsType | undefined) => {
   const { resumeTask, checkTask, fetchTaskDetails } = useTaskDetails()
   const [isResuming, setIsResuming] = useState(false)
   const [isChecking, setIsChecking] = useState(false)
@@ -17,7 +17,7 @@ export const useTaskActions = (task: TaskDetailsType | undefined) => {
       const success = await resumeTask(task.id)
       if (success) {
         void fetchTaskDetails(task.id)
-        toast.success('Задача возобновлена')
+        toast.success('Р—Р°РґР°С‡Р° РІРѕР·РѕР±РЅРѕРІР»РµРЅР°')
       }
     } finally {
       setIsResuming(false)
@@ -31,7 +31,7 @@ export const useTaskActions = (task: TaskDetailsType | undefined) => {
       const success = await checkTask(task.id)
       if (success) {
         void fetchTaskDetails(task.id)
-        toast.success('Задача проверена')
+        toast.success('Р—Р°РґР°С‡Р° РїСЂРѕРІРµСЂРµРЅР°')
       }
     } finally {
       setIsChecking(false)
@@ -40,3 +40,5 @@ export const useTaskActions = (task: TaskDetailsType | undefined) => {
 
   return { isResuming, isChecking, canResume, handleResume, handleCheck }
 }
+
+export const useTaskActions = useTaskDetailsActions
